fix pr for comment
